fix battery percentage detection
The last byte is actually the frame counter.

This is clearly visible when plotted and explains the constantly
changing battery values that went over 100% ;-)

Still unsure if humidity has 2 bytes (the plot looked good before) or
if the second byte is the battery percentage as stated by the
documentation.  Let's change it, plot it and see.